Remove auto-update

+++ 04-Extras.sh
@@ -41,6 +41,13 @@ error_msg() {
 
 cd "${TOOLKIT_PATH}"
 
+if [[ "$(uname -m)" != "x86_64" ]]; then
+  echo "Error: This script requires an x86-64 CPU architecture. Detected: $(uname -m)" | tee -a "${LOG_FILE}"
+  read -n 1 -s -r -p "Press any key to exit..."
+  echo
+  exit 1
+fi
+
 echo "########################################################################################################" | tee -a "${LOG_FILE}" >/dev/null 2>&1
 if [ $? -ne 0 ]; then
     sudo rm -f "${LOG_FILE}"
@@ -54,55 +61,12 @@ if [ $? -ne 0 ]; then
     fi
 fi
 
-# Check if the current directory is a Git repository
-if ! git rev-parse --is-inside-work-tree > /dev/null 2>&1; then
-  echo "This is not a Git repository. Skipping update check." | tee -a "${LOG_FILE}"
-else
-  # Fetch updates from the remote
-  git fetch >> "${LOG_FILE}" 2>&1
-
-  # Check the current status of the repository
-  LOCAL=$(git rev-parse @)
-  REMOTE=$(git rev-parse @{u})
-  BASE=$(git merge-base @ @{u})
-
-  if [ "$LOCAL" = "$REMOTE" ]; then
-    echo "The repository is up to date." | tee -a "${LOG_FILE}"
-  else
-    echo "Downloading updates..."
-    # Get a list of files that have changed remotely
-    UPDATED_FILES=$(git diff --name-only "$LOCAL" "$REMOTE")
-
-    if [ -n "$UPDATED_FILES" ]; then
-      echo "Files updated in the remote repository:" | tee -a "${LOG_FILE}"
-      echo "$UPDATED_FILES" | tee -a "${LOG_FILE}"
-
-      # Reset only the files that were updated remotely (discard local changes to them)
-      echo "$UPDATED_FILES" | xargs git checkout -- >> "${LOG_FILE}" 2>&1
-
-      # Pull the latest changes
-      if ! git pull --ff-only >> "${LOG_FILE}" 2>&1; then
-        error_msg "Error" "Update failed. Delete the PSBBN-Definitive-English-Patch directory and run the command:" " " "git clone https://github.com/CosmicScale/PSBBN-Definitive-English-Patch.git" " " "Then try running the script again."
-      fi
-      echo
-      echo "The repository has been successfully updated." | tee -a "${LOG_FILE}"
-      read -n 1 -s -r -p "Press any key to exit, then run the script again."
-      echo
-      exit 0
-    fi
-  fi
-fi
-
 date >> "${LOG_FILE}"
 echo >> "${LOG_FILE}"
 cat /etc/*-release >> "${LOG_FILE}" 2>&1
-
-if [[ "$(uname -m)" != "x86_64" ]]; then
-  echo "Error: This script requires an x86-64 CPU architecture. Detected: $(uname -m)" | tee -a "${LOG_FILE}"
-  read -n 1 -s -r -p "Press any key to exit..."
-  echo
-  exit 1
-fi
+echo >> "${LOG_FILE}"
+echo "PSBBN Definitive Patch Version 2.11" >> "${LOG_FILE}"
+echo "Path set to: $TOOLKIT_PATH" >> "${LOG_FILE}"
 
 # Function to detect PS2 HDD
